Move context files to proper directories
- Moved build_npc_context.py to core/ai/
- Moved inventory_context_integration.py to core/ai/
- Moved inventory_context_matcher.py to core/ai/
- Moved inventory_context_matcher_v2.py to core/ai/
- Created tests/ directory and moved test_validation_npc_context.py
- Updated all imports in main.py and combat_manager.py
- Fixed internal import in inventory_context_integration.py
