+++ core/ai/build_npc_context.py
