+++ core/ai/inventory_context_matcher.py
