+++ core/ai/inventory_context_matcher_v2.py
